% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{bestplane}
\alias{bestplane}
\title{Calculate the least squares plane of some points}
\usage{
bestplane(l)
}
\arguments{
\item{l}{An n x 3 matrix of landmarks to be used in creating the plane}
}
\value{
A list containing a unit vector n and a scalar d.
 n.x = d is the least squares plane.
}
\description{
Calculate the least squares plane of some points
}

