% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countMissing}
\alias{countMissing}
\title{Count the number of missing landmarks in an array}
\usage{
countMissing(A)
}
\arguments{
\item{A}{An n x m x 3 array where n is the number of specimens and m is the number of landmarks.}
}
\value{
A length n vector giving the number of missing landmarks for each specimen.
}
\description{
Count the number of missing landmarks in an array
}
\examples{



 a <- array(1:(3*6*7), dim = c(7, 3, 6))
 a[2, , 1] <- NA
 countMissing(a)
}

