% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{cvm}
\alias{cvm}
\title{Calculate 1D covariance between two landmarks across specimens}
\usage{
cvm(M)
}
\arguments{
\item{M}{An M x N matrix. M = no of specimens, N = no of landmarks.}
}
\value{
1D covariance matrix
}
\description{
Calculate 1D covariance between two landmarks across specimens
}

