% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{deltaa}
\alias{deltaa}
\title{Returns the sum of squares of the distances between "a1" and "a2".}
\usage{
deltaa(olda, newa, m, n, scaleDelta, zap = TRUE)
}
\arguments{
\item{olda}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}

\item{newa}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}

\item{m}{No of specimens}

\item{n}{No of landmarks}
}
\value{
The sum of squares distances (length 1 numeric) between all landmarks on all speciments.
}
\description{
For each landmark on each sample, find distance between location given
  in a1 and a2. 
Used to see when a1 and a2 are very similar. e.g. deltaa(olda, newa, 10, 20, scaleDelta = FALSE) < 10e-7
}

