% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{dotcorrentry}
\alias{dotcorrentry}
\title{Calculate 3D correlation between two landmarks across specimens}
\usage{
dotcorrentry(M, col1, col2)
}
\arguments{
\item{M}{An N x 3 x M array. M = no of specimens, N = no of landmarks.}

\item{col1}{Integer of first column}

\item{col2}{Integer of second column}
}
\value{
Length 1 numeric giving the congruence correlation
}
\description{
Calculates the congruence coefficient for 2 or 3 dimensional landmarks
}

