% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{dotcovar}
\alias{dotcovar}
\title{Calculate covariance matrix between individual dimensions within landmarks}
\usage{
dotcovar(M)
}
\arguments{
\item{M}{An N x 3 x M array. M = no of specimens, N = no of landmarks.}
}
\value{
3N x 3N covariance matrix
}
\description{
Calculate covariance matrix between individual dimensions within landmarks. Skips any missing values
  in computation of covariance matrix.
}
\details{
This function does not guarantee that the returned matrix is  
 positive definite. If the covariance matrix is not positive definite 
 a warning is given and the matrix can be bent to create the closest
 positive definite matrix with \code{as.matrix(Matrix::nearPD(mat)$mat)}.
}
\examples{
M <- array(rnorm(4 * 2 * 3), dim = c(2, 3, 4)) 
M.cvm <- dotcovar(M)
}

