% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{dotcvmd}
\alias{dotcvm}
\alias{dotcvmd}
\title{Calculate 3D covariance matrix}
\usage{
dotcvm(M)
}
\arguments{
\item{M}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}
}
\value{
N x N covariance matrix
}
\description{
Calculate 3D covariance matrix
}

