% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{dotcvmentry}
\alias{dotcvmentry}
\title{Calculate 3D covariance between two landmarks across specimens}
\usage{
dotcvmentry(M, col1, col2)
}
\arguments{
\item{M}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}

\item{col1}{Integer of first column}

\item{col2}{Integer of second column}
}
\value{
Covariance value
}
\description{
Calculate 3D covariance between two landmarks across specimens
}

