% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMLi.R
\name{EMMLi}
\alias{EMMLi}
\title{Evaluating modularity with maximum likelihood}
\usage{
EMMLi(corr, N_sample, mod, saveAs)
}
\arguments{
\item{corr}{Lower triangle or full correlation matrix. n x n square matrix for n landmarks.}

\item{N_sample}{The number of samples used}

\item{mod}{A data frame defining the models. The first column should contain the landmark names. Subsequent columns should define which landmarks are contained with which module. If a landmark should be ignored for a specific model, the element should be NA.}

\item{saveAs}{A character string defining the filename and path for where to save output.}
}
\value{
NULL. The output is saved to the file defined by the saveAs argument.
}
\description{
Calculates the AICc values of different models of modularity.
}
\examples{
 set.seed(1)

 # Chose a filename and directory for output
 file <- paste0(tempdir(), 'EMMLiTest.csv')
 
 # Create a correlation matrix
 dat <- matrix(runif(36, -1, 1), ncol = 6, nrow = 6)
 diag(dat) <- 1

 # Create dataframe of model definitions.
 #   Use NAs indicate landmarks which are not part of the model
 mod1 <- data.frame(landmarks = letters[1:6], 
              modelA = rep(c(1, 2), each = 3),
              modelB = rep(c(1,2), 3),
              model3= c(rep(c(1,2), 2), NA, NA)) 


 # run EMMLi
 EMMLi(dat, 20, mod1, file)

 unlink(file)
}

