% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{lshift}
\alias{lcentroid}
\alias{lcentroid2}
\alias{lnorm}
\alias{lrotate}
\alias{lscale}
\alias{lshift}
\title{Common manipulations of 3D points}
\usage{
lshift(A, v)

lrotate(A, m)

lcentroid(A)

lcentroid2(A)

lnorm(A)

lscale(A, v)
}
\arguments{
\item{A}{An n x 3 matrix}

\item{v}{Length 3 displacement vector}

\item{m}{3 x 3 rotation matrix}

\item{A}{An n x 3 matrix}

\item{A}{An n x 3 matrix}

\item{A}{An n x 3 matrix}

\item{A}{An n x 3 matrix}

\item{A}{An n x 3 matrix}

\item{v}{Length 3 scaling vector}
}
\value{
A matrix of the same dimensions as A

A matrix of the same dimensions as A

A matrix of the same dimensions as A

A matrix of the same dimensions as A

A scalar of total square length.

A scalar of total square length.
}
\description{
lshift: translates all 3D points in matrix A by a common displacement

lrotate: rotates all 3D points in matrix A by a common matrix

lcentroid: computes centroid of all 3D points in matrix A (cannot have missing data)

lcentroid2: like lcentroid but tolerates missing data

lnorm: returns total square length, tolerates missing data

lscale: like lshift[], but multiplicative not additive
}

