% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{midline}
\alias{midline}
\title{Find the midline of an object}
\usage{
midline(s, l1)
}
\arguments{
\item{s}{An n x 3 matrix of the 3D coordinates of n landmarks.}

\item{l1}{Indices of the landmarks to use in creating the midline.}
}
\description{
Find the midline of an object
}

