% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirrorfill1}
\alias{mirrorfill1}
\title{Fill missing landmarks using mirrored values from other side of object}
\usage{
mirrorfill1(s, l1, l2)
}
\arguments{
\item{s}{An n x 3 matrix containing 3D landmark data of n landmarks.}

\item{l1}{Vector of indices for which landmarks to use to make a specimen midline.}

\item{l2}{Vector of indices for which landmarks to be replaced by their mirrored value.}
}
\description{
Given an n x 3 matrix, replace a set of landmarks using their mirrored counterpark.
}
\details{
\code{l2} should be an even number length containing pairs of landmarks
 on either side of the specimen. i.e. l2[1]
}

