% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{pcrstep}
\alias{pcrstep}
\title{Rotate all shapes until optimally aligned.}
\usage{
pcrstep(a, maxiter = 10000, tolerance = 1e-06)
}
\arguments{
\item{a}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}

\item{maxiter}{Maximum number of iterations to attempt}

\item{tolerance}{Difference between two iterations that will cause the search to stop.}
}
\value{
An M x N x 3 array of aligned shapes
}
\description{
Given an M x N x 3 array (M = no of specimens, N = no of landmarks.) 
  this will find the optimal rotation for all shapes so that they are as
  aligned as possible.
}

