% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{pctstep}
\alias{pctstep}
\title{Translate all shapes until optimally aligned.}
\usage{
pctstep(a)
}
\arguments{
\item{a}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}
}
\value{
An M x N x 3 array of aligned shapes
}
\description{
Given an M x N x 3 array (M = no of specimens, N = no of landmarks.) 
  this will find the optimal translation for all shapes so that they are as
  aligned as possible.
}

