% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpecimens.R
\name{plotSpecimens}
\alias{plotSpecimens}
\title{Plot an array of specimen landmark data in an interactive 3D frame}
\usage{
plotSpecimens(a, cols = NULL, bySpecimen = TRUE, ...)
}
\arguments{
\item{a}{An M x N x 3 array.}

\item{cols}{A vector of colours.}

\item{...}{Further parameters passed to \code{plot3d}.}

\item{bylandmark}{Logical that determined whether points should be coloured by specimen (default) or by landmark.}
}
\description{
This function requires the rgl package.
  Given a M x N x 3 array (where M is the number of speciments
  and N is the number of landmarks), as used elsewhere in this package,
  plot each specimen in a different colour in an intereactive
  3D frame.
}
\examples{
a <- array(rep(rnorm(6 * 20, sd = 30), each = 6) + rnorm(6 * 20 * 3), 
       dim = c(6, 20, 3))
plotSpecimens(a)

plotSpecimens(a, bySpecimen = FALSE)

plotSpecimens(a, cols = grey(seq(0, 1, length.out = 6)))


}
\seealso{
\code{\link[rgl]{plot3d}}
}

