% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustespkg.R
\docType{package}
\name{procrustespkg}
\alias{procrustespkg}
\alias{procrustespkg-package}
\title{procrustespkg: Proctustes Analysis for Paleobiology}
\description{
procrustespkg: Proctustes Analysis for Paleobiology
}

