% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{reflect}
\alias{reflect}
\title{Reflects p in the plane defined by n.x = d
n is a unit vector. d is a scalar.}
\usage{
reflect(p, n, d)
}
\arguments{
\item{p}{a length 3 vector (a 3D point in space)}

\item{n}{Length three vector of coefficients for a plane nx = d}

\item{d}{Length one vector giving coefficient d for a plane nx = d}
}
\description{
Reflects p in the plane defined by n.x = d
n is a unit vector. d is a scalar.
}

