% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{rp_decompose}
\alias{rp_decompose}
\title{Decompose matric to a rotation and symetric positive definite matrix}
\usage{
rp_decompose(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A list with matrices R (the rotation) and P (the symetric 
 positive definite matrix) such that M = RP.
}
\description{
rpdecompose: Any invertible matrix can be uniquely decomposed
  M = RP, where R is a rotation and P is symmetric positive
  definite.  This returns the list {R,P}.
}

