% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{scorea}
\alias{scorea}
\title{Returns the sum of squares of distances that we're trying to minimize.}
\usage{
scorea(arr, m, n)
}
\arguments{
\item{arr}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}

\item{m}{No of specimens}

\item{n}{No of landmarks}
}
\value{
The sum of squares distance
}
\description{
Returns the sum of squares of distances that we're trying to minimize.
}

