% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{unzapa}
\alias{unzapa}
\title{Puts missing data back in to a specimen x landmark array}
\usage{
unzapa(a, b)
}
\arguments{
\item{a}{An M x N x 3 array.}

\item{b}{An M x N x 3 array to use a template.}
}
\value{
An M x N x 3 array with NAs for missing data.
}
\description{
Given an M x N x 3 array, and a template defining which data were
  missing, returns an M x N x 3 array with NAs for missing data.
  M is the number of specimens and N is the number of landmarks.
}

