% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{zapa}
\alias{zapa}
\title{Replaces missing data points with c(0, 0, 0)}
\usage{
zapa(a)
}
\arguments{
\item{a}{An M x N x 3 array.}
}
\value{
An M x N x 3 array with no missing data.
}
\description{
Given an M x N x 3 array, returns an M x N x 3 array with no missing data.
  M is the number of specimens and N is the number of landmarks.
}

