% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{dotcorr}
\alias{dotcorr}
\title{Calculate 3D correlation matrix}
\usage{
dotcorr(M)
}
\arguments{
\item{M}{An M x N x D array. M = no of specimens, N = no of landmarks, D = 2 or 3 dimensions}
}
\value{
Correlation matrix
}
\description{
Calculates the congruence coefficient for 2 or 3 dimensional landmarks
  to give a M x M correlation matrix.
}
\examples{
M <- array(rnorm(4 * 2 * 3), dim = c(4, 2, 3)) 
M.corr <- dotcorr(M)

}

