% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{dotcorr}
\alias{dotcorr}
\title{Calculate 3D correlation matrix}
\usage{
dotcorr(M)
}
\arguments{
\item{M}{An N x D x M array. M = no of specimens, N = no of landmarks, D = 3 dimensions}
}
\value{
Correlation matrix
}
\description{
Calculates the congruence coefficient for 3 dimensional landmarks
  to give an N x N correlation matrix.
}
\examples{
M <- array(rnorm(4 * 2 * 3), dim = c(2, 3, 4)) 
M.corr <- dotcorr(M)

}

