% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{dotcvm}
\alias{dotcvm}
\title{Calculate 3D covariance matrix}
\usage{
dotcvm(M)
}
\arguments{
\item{M}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}
}
\value{
N x N covariance matrix
}
\description{
Calculate 3D covariance matrix
}
\details{
This function does not guarantee that the returned matrix is  
 positive definite. If the covariance matrix is not positive definite 
 a warning is given and the matrix can be bent to create the closest
 positive definite matrix with \code{as.matrix(Matrix::nearPD(mat)$mat)}.
}

