% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{procrustes}
\alias{procrustes}
\title{Run procrustese analysis to align 3D shapes.}
\usage{
procrustes(a, scale = TRUE, maxiter = 1000, tolerance = 1e-05)
}
\arguments{
\item{a}{An M x N x 3 array. M = no of specimens, N = no of landmarks.}

\item{scale}{Logical indicating whether the size of the objects should be scales
as well as rotated and translated.}

\item{maxiter}{Maximum number of iterations to attempt}

\item{tolerance}{Difference between two iterations that will cause the search to stop.}
}
\value{
A new (M x N x 3) array, where each 3d vector has been transformed
 in a per-specimen way.  The transformation is chosen to maximize,
 in the least-squares sense, the distances between specimens.
}
\description{
Run procrustese analysis to align 3D shapes.
}

